requirements.txt
django==4.0.5
django-debug-toolbar==3.2.1
Pillow==9.1.0
environs[django]==9.3.2
djangorestframework==3.13.1
django-phonenumber-field==6.1.0
phonenumbers==8.12.48
requests==2.27.1
geopy==2.2.0
django-admin-sortable2==2.1
